Añadidas comprobaciones de rol al hacer crud sobre orders y products

+++ src/api/controllers/order.js
@@ -1,4 +1,5 @@
 const Order = require("../models/Order");
+const User = require("../models/User");
 
 const getOrder = async (req, res, next) => {
   try {
@@ -22,8 +23,12 @@ const getOrderById = async (req, res, next) => {
 const addOrder = async (req, res, next) => {
   try {
     const order = new Order(req.body);
-
     const orderBD = await order.save();
+
+    const user = await User.findById(req.user._id);
+    user.orders.push(orderBD._id);
+    await User.findByIdAndUpdate(user._id, user);
+
     return res.status(201).json(orderBD);
   } catch (error) {
     next(error);
@@ -33,9 +38,16 @@ const addOrder = async (req, res, next) => {
 const updateOrder = async (req, res, next) => {
   try {
     const { id } = req.params;
-
+    const order = await Order.findById(id).populate("products");
     const newOrder = new Order(req.body);
     newOrder._id = id;
+
+    // Se comprueba que no se repitan los productos
+    order.products.forEach((prod) => {
+      if (!newOrder.products.includes(prod._id)) {
+        newOrder.products.push(prod);
+      }
+    });
 
     const updatedOrder = await Order.findByIdAndUpdate(id, newOrder, {
       new: true,
@@ -49,6 +61,11 @@ const updateOrder = async (req, res, next) => {
 const deleteOrder = async (req, res, next) => {
   try {
     const { id } = req.params;
+
+    if (req.user.role == "user") {
+      return res.status(401).json("Only admins can delete orders");
+    }
+
     await Order.findByIdAndDelete(id);
     return res.status(200).json("Order deleted");
   } catch (error) {
